# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BondCircle Mobile API
  version: 1.0.0
  description: Official versioned REST API contract for BondCircle native Android and web clients.

servers:
  - url: https://bondcircle.app
    description: Production Server
  - url: https://staging.bondcircle.app
    description: Staging Server
  - url: http://10.0.2.2:3000
    description: Local Android Emulator

security:
  - BearerAuth: []

paths:
  /api/v1/me:
    get:
      summary: Bootstrap current authenticated user session
      description: Returns the user profile, legal acceptance timestamps, and notification preferences.
      responses:
        "200":
          description: User profile payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Authentication required

  /api/v1/auth/email-otp/request:
    post:
      summary: Request email OTP challenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailOtpRequest"
      responses:
        "202":
          description: Challenge issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailOtpResponse"
        "400":
          description: Validation error

  /api/v1/auth/email-otp/verify:
    post:
      summary: Verify email OTP code and receive Firebase Custom Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailOtpRequest"
      responses:
        "200":
          description: Verification success with custom token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyEmailOtpResponse"
        "400":
          description: Invalid or expired code

  /api/v1/circles:
    get:
      summary: List user circles
      description: Returns paginated circles where the user is creator, co-admin, or member, with privacy redaction applied.
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Circle list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CircleListResponse"

  /api/v1/devices/push-token:
    post:
      summary: Register FCM device token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevicePushTokenRequest"
      responses:
        "200":
          description: Token registered
    delete:
      summary: Remove FCM device token
      responses:
        "200":
          description: Token removed

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserProfileResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"

    UserProfile:
      type: object
      required:
        - id
        - displayName
        - accountStatus
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        profileImage:
          type: string
          nullable: true
        termsAcceptedAt:
          type: string
          nullable: true
        privacyAcceptedAt:
          type: string
          nullable: true
        accountStatus:
          type: string
          enum: [active, suspended]
        preferences:
          $ref: "#/components/schemas/UserPreferences"

    UserPreferences:
      type: object
      properties:
        emailNotifications:
          type: boolean
        browserPushNotifications:
          type: boolean
        commentNotifications:
          type: boolean
        contributionReminders:
          type: boolean
        circleUpdateNotifications:
          type: boolean
        marketingCommunication:
          type: boolean

    EmailOtpRequest:
      type: object
      required:
        - email
        - purpose
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum: [sign-in, register, recovery]
        displayName:
          type: string
        termsAccepted:
          type: boolean
        privacyAccepted:
          type: boolean

    EmailOtpResponse:
      type: object
      required:
        - challengeId
        - expiresAt
        - message
      properties:
        challengeId:
          type: string
        expiresAt:
          type: string
        message:
          type: string
        delivery:
          type: string
        developmentCode:
          type: string

    VerifyEmailOtpRequest:
      type: object
      required:
        - challengeId
        - code
      properties:
        challengeId:
          type: string
        code:
          type: string

    VerifyEmailOtpResponse:
      type: object
      required:
        - customToken
        - uid
      properties:
        customToken:
          type: string
        uid:
          type: string

    CircleSummary:
      type: object
      required:
        - id
        - name
        - type
        - status
        - memberCount
        - memberLimit
        - role
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [GIFT, ASO_EBI, SUPPORT]
        imageUrl:
          type: string
          nullable: true
        targetAmount:
          type: integer
          nullable: true
        currentAmount:
          type: integer
        memberCount:
          type: integer
        memberLimit:
          type: integer
        status:
          type: string
          enum: [DRAFT, ACTIVE, COMPLETED, CANCELLED, ARCHIVED]
        deadline:
          type: string
          nullable: true
        role:
          type: string
          enum: [creator, co_admin, member]
        createdAt:
          type: string

    CircleListResponse:
      type: object
      required:
        - circles
      properties:
        circles:
          type: array
          items:
            $ref: "#/components/schemas/CircleSummary"
        nextCursor:
          type: string
          nullable: true

    DevicePushTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
        platform:
          type: string
          default: android